requirements.txt
# Core dependencies
chainlit>=1.0.0
mistralai>=1.9.1
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0

# Memory layer for personalization
mem0ai>=0.1.0

# Web search
exa-py>=1.14.0

# Date/time utilities
python-dateutil>=2.8.0

# Math calculations (for BMI, workout planning)
numpy>=1.24.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0 